+++ core-service/src/main/java/ru/gb/shop/core/entities/OrderItem.java
@@ -29,10 +29,10 @@ public class OrderItem {
     private int quantity;
 
     @Column(name = "price_per_product")
-    private int pricePerProduct;
+    private double pricePerProduct;
 
     @Column(name = "price")
-    private int price;
+    private double price;
 
     @CreationTimestamp
     @Column(name = "created_at")
@@ -41,7 +41,7 @@ public class OrderItem {
     @Column(name = "updated_at")
     private LocalDateTime updatedAt;
 
-    public OrderItem(Product product, Order order, int quantity, int pricePerProduct, int price) {
+    public OrderItem(Product product, Order order, int quantity, double pricePerProduct, double price) {
         this.product = product;
         this.order = order;
         this.quantity = quantity;
